+++ sync-for-payroll/src/sdk/models/shared/createaccountresponse.ts
@@ -143,7 +143,7 @@ export class CreateAccountResponse extends SpeakeasyBase {
     requestedOnUtc: string;
 
     /**
-     * The status of the push operation.
+     * The current status of the push operation.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "status" })